refactor: move constants and decorators to core package
