+++ kiauh/components/klipper_firmware/menus/klipper_build_menu.py
@@ -17,9 +17,9 @@ from components.klipper_firmware.firmware_utils import (
     run_make_clean,
     run_make_menuconfig,
 )
+from core.constants import COLOR_CYAN, COLOR_GREEN, COLOR_RED, RESET_FORMAT
 from core.menus import Option
 from core.menus.base_menu import BaseMenu
-from utils.constants import COLOR_CYAN, COLOR_GREEN, COLOR_RED, RESET_FORMAT
 from utils.logger import Logger
 from utils.sys_utils import (
     check_package_install,
