+++ kiauh/components/mobileraker/__init__.py
@@ -9,7 +9,7 @@
 from pathlib import Path
 
 from core.backup_manager import BACKUP_ROOT_DIR
-from utils.constants import SYSTEMD
+from core.constants import SYSTEMD
 
 # repo
 MOBILERAKER_REPO = "https://github.com/Clon1998/mobileraker_companion.git"
