+++ kiauh/components/webui_client/menus/client_remove_menu.py
@@ -13,9 +13,9 @@ from typing import Type
 
 from components.webui_client import client_remove
 from components.webui_client.base_data import BaseWebClient, WebClientType
+from core.constants import COLOR_CYAN, COLOR_RED, RESET_FORMAT
 from core.menus import Option
 from core.menus.base_menu import BaseMenu
-from utils.constants import COLOR_CYAN, COLOR_RED, RESET_FORMAT
 
 
 # noinspection PyUnusedLocal
