+++ kiauh/utils/common.py
@@ -14,13 +14,13 @@ from pathlib import Path
 from typing import Dict, List, Literal, Optional, Type
 
 from components.klipper.klipper import Klipper
-from core.instance_manager.base_instance import BaseInstance
-from core.instance_manager.instance_manager import InstanceManager
-from utils import GLOBAL_DEPS, PRINTER_CFG_BACKUP_DIR
-from utils.constants import (
+from core.constants import (
     COLOR_CYAN,
     RESET_FORMAT,
 )
+from core.instance_manager.base_instance import BaseInstance
+from core.instance_manager.instance_manager import InstanceManager
+from utils import GLOBAL_DEPS, PRINTER_CFG_BACKUP_DIR
 from utils.git_utils import get_local_commit, get_remote_commit, get_repo_name
 from utils.logger import DialogType, Logger
 from utils.sys_utils import (
